refactor(core): extract recall-rerank seam from orchestrator (#1526 seam 10)
