+++ scripts/ratchet-baseline.json
@@ -4,7 +4,7 @@
   "oversizeThresholdLoc": 3000,
   "metrics": {
     "watchlistLoc": {
-      "packages/remnic-core/src/orchestrator.ts": 19261,
+      "packages/remnic-core/src/orchestrator.ts": 18940,
       "packages/remnic-core/src/cli.ts": 9393,
       "packages/remnic-core/src/access-service.ts": 9016,
       "packages/remnic-core/src/storage.ts": 7747,
